+++ src/osm_configurator/model/parser/custom_expceptions/osm_data_wrongly_formatted_Exception.py
