+++ src/osm_configurator/model/project/calculation/tag_filter_phase.py
@@ -11,7 +11,7 @@ import src.osm_configurator.model.project.calculation.calculation_phase_utility 
 
 from src.osm_configurator.model.project.calculation.calculation_phase_interface import ICalculationPhase
 from osm_configurator.model.parser.custom_expceptions.tags_wrongly_formatted_exception import TagsWronglyFormatted
-from osm_configurator.model.parser.custom_expceptions.osm_data_wrongly_formatted import OSMDataWronglyFormatted
+from osm_configurator.model.parser.custom_expceptions.osm_data_wrongly_formatted_Exception import OSMDataWronglyFormatted
 from src.osm_configurator.model.parser.custom_expceptions.illegal_cut_out_exception import IllegalCutOutException
 import src.osm_configurator.model.parser.cut_out_parser as cut_out_parser
